Refactor timer_module
fixed method order and moved match ending stuff to a new file
